Create testing_index.cgi
With this, the automated column count of table headers is now present, and full columns can now be hidden/shown. I still have to go thru editing the administrative side for changing the ability to show/hide columns, so this is not ready for full integration into the set of scripts, but available if desired to show the progress.

I did have to change the first line a bit, as I didn't realize that having the date in the second entry would cause issues for the column show/hide ability. So, rather than "#DATE#|2017.02.18|", this will now be "#DATE#,2017.02.18|".

To show/hide columns, currently you have to manually edit the first line of the database files to be "#DATE#,2017.02.18|" followed the addition of "show|" to display columns or "hide|" to hide columns. This will be integrated into the administrative side soon.