Update media.pl
On the videos section, I had included an if/then statement that would change the media type of entries to "Streaming" if there wasn't a media type option selected. This revision modifies this slightly, in that "Streaming" will only show if a streaming service is selected, otherwise it won't show anything.