Update index.cgi
I found a bug I introduced in which having "mobile information" enabled would cause the videos section to most likely show a 500 error page on the server. This is due to a number of repeated "<BR>" entries which would show up at the top of the script and this would be interpreted as a CGI header. When I introduced the "mobile information" option, I didn't test all possibilities, and discovered this. 2017.09.18 update to the index.cgi file corrects this problem, as well as a little bit of code cleanup.